Split the types into separate files
