+++ src/macro_aliases.rs
@@ -1,0 +1,5 @@
+use macro_rules_attribute::derive_alias;
+
+derive_alias! {
+    #[derive(Serde!)] = #[derive(Deserialize, Clone, Debug)];
+}
