+++ src/trait_implementations/timeline.rs
@@ -6,7 +6,7 @@ use crate::{
 impl Timeline for Edupage {
     fn filter_timeline_by_item_type(
         &self,
-        item_type: crate::edupage_types::TimelineItemType,
+        item_type: crate::edupage_types::timeline::TimelineItemType,
     ) -> Result<Vec<crate::edupage_types::TimelineItem>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
@@ -26,7 +26,7 @@ impl Timeline for Edupage {
 
     fn filter_timeline_by_item_types(
         &self,
-        item_types: Vec<crate::edupage_types::TimelineItemType>,
+        item_types: Vec<crate::edupage_types::timeline::TimelineItemType>,
     ) -> Result<Vec<crate::edupage_types::TimelineItem>, crate::edupage::EdupageError> {
         if !self.is_logged_in {
             return Err(EdupageError::NotLoggedIn);
